+++ tests/evaluation/run/test_runner.py
@@ -215,3 +215,23 @@ def test_runner_run_overview_has_specified_metadata_and_labels(
 
     assert overview.metadata == run_metadata
     assert overview.labels == run_labels
+
+
+def test_run_is_already_computed_works(
+    in_memory_dataset_repository: InMemoryDatasetRepository,
+    in_memory_run_repository: InMemoryRunRepository,
+    sequence_examples: Iterable[Example[str, None]],
+) -> None:
+    old_model = "old_model"
+    examples = list(sequence_examples)
+    task = DummyTask()
+    runner = Runner(task, in_memory_dataset_repository, in_memory_run_repository, "foo")
+    dataset_id = in_memory_dataset_repository.create_dataset(
+        examples=examples, dataset_name=""
+    ).id
+
+    run_metadata: SerializableDict = dict({"model": old_model})
+    runner.run_dataset(dataset_id, metadata=run_metadata)
+
+    assert runner.run_is_already_computed(dict({"model": old_model}))
+    assert not runner.run_is_already_computed(dict({"model": "new_model"}))