+++ CHANGELOG.md
@@ -3,12 +3,24 @@
 
 ### Features
 ...
+
 ### Fixes
 ...
+
 ### Deprecations 
 ...
 ### Breaking Changes
 ...
+
+## 7.1.0
+
+### Features
+- Add Catalan and Polish support to `DetectLanguage`.
+- Add utility function `run_is_already_computed` to `Runner` to check if a run with the given metadata has already been computed.
+  - The `parameter_optimization` notebook describes how to use the `run_is_already_computed` function.
+  
+### Fixes
+- The default `max_retry_time` for the `LimitedConcurrencyClient` is now set to 3 minutes from a day. If you have long-running evaluations that need this, you can re-set a long retry time in the constructor.
 
 
 ## 7.0.0
@@ -24,8 +36,6 @@
 ### Fixes
 - `DocumentIndexClient` now properly sets `chunk_overlap` when creating an index configuration.
 
-### Deprecations 
-...
 ### Breaking Changes
 - The default model for `Llama3InstructModel` is now `llama-3.1-8b-instruct` instead of `llama-3-8b-instruct`. We also removed the llama3.0 models from the recommended models of the `Llama3InstructModel`.
 - The default value of `threshold` in the `DocumentIndexRetriever` has changed from `0.5` to `0.0`. This accommodates fusion scoring for searches over hybrid indexes.